requirements.txt
streamlit
scikit-learn
Pillow
opencv-python-headless
tensorflow==2.11.0
numpy<1.24